requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
blinker==1.4
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
cryptography==2.3.1
defusedxml==0.5.0
Django==2.1.2
django-allauth==0.38.0
future==0.17.0
gpiozero==1.4.1
httplib2==0.9.2
hyperlink==17.3.1
idna==2.7
incremental==16.10.1
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
MarkupSafe==1.0
oauthlib==2.1.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.19
pycrypto==2.6.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyserial==3.4
python-debian==0.1.32
python-telegram-bot==11.1.0
python3-openid==3.1.0
pytz==2018.6
pyxdg==0.25
PyYAML==3.12
requests==2.20.0
requests-oauthlib==1.0.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==16.0.0
six==1.11.0
ssh-import-id==5.6
systemd-python==234
Twisted==17.9.0
urllib3==1.24
waitress==1.1.0
zope.interface==4.3.2
